+++ includes/widgets/w-subnav.php
@@ -6,7 +6,7 @@
  * @subpackage  Widgets
  *
  * @since    1.0.9.9
- * @version  1.2.8
+ * @version  1.3.2
  */
 
 
@@ -48,18 +48,17 @@ class WM_Subnav extends WP_Widget {
 
 		/**
 		 * Constructor
-		 *
-		 * @since    1.0.9.9
-		 * @version  1.2.8
 		 */
 		function __construct() {
 
 			// Helper variables
 
+				$theme = ( is_child_theme() ) ? ( wp_get_theme()->parent()->get_template() ) : ( null );
+
 				$atts = array();
 
 				$atts['id']          = 'wm-subnav';
-				$atts['name']        = wp_get_theme()->get( 'Name' ) . ' ' . esc_html_x( 'Submenu', 'Widget name.', 'webman-amplifier' );
+				$atts['name']        = wp_get_theme( $theme )->get( 'Name' ) . ' ' . esc_html_x( 'Submenu', 'Widget name.', 'webman-amplifier' );
 				$atts['widget_ops']  = array(
 						'classname'   => 'wm-subnav',
 						'description' => esc_html_x( 'List of subpages', 'Widget description.', 'webman-amplifier' )
@@ -176,7 +175,7 @@ class WM_Subnav extends WP_Widget {
 				// Title
 
 					if ( trim( $instance['title'] ) ) {
-						$output .= $args['before_title'] . apply_filters( 'widget_title', $instance['title'] ) . $args['after_title'];
+						$output .= $args['before_title'] . apply_filters( 'widget_title', $instance['title'], $instance, $this->id_base, $args ) . $args['after_title'];
 					}
 
 					$output .= '<ul class="sub-nav">' . $children . '</ul>';