+++ includes/widgets/w-twitter.php
@@ -8,7 +8,7 @@
  * @subpackage  Widgets
  *
  * @since    1.0.9.9
- * @version  1.2.2
+ * @version  1.3.2
  *
  * CONTENT:
  * - 10) Actions and filters
@@ -85,10 +85,12 @@ if ( ! defined( 'ABSPATH' ) ) exit;
 
 			// Helper variables
 
+				$theme = ( is_child_theme() ) ? ( wp_get_theme()->parent()->get_template() ) : ( null );
+
 				$atts = array();
 
 				$atts['id']          = 'wm-twitter';
-				$atts['name']        = wp_get_theme()->get( 'Name' ) . ' ' . esc_html_x( 'Twitter', 'Widget name.', 'webman-amplifier' );
+				$atts['name']        = wp_get_theme( $theme )->get( 'Name' ) . ' ' . esc_html_x( 'Twitter', 'Widget name.', 'webman-amplifier' );
 				$atts['widget_ops']  = array(
 						'classname'   => 'wm-twitter',
 						'description' => _x( 'Your recent tweets', 'Widget description.', 'webman-amplifier' )
@@ -390,7 +392,7 @@ if ( ! defined( 'ABSPATH' ) ) exit;
 					$output .= $args['before_widget'];
 
 					if ( trim( $instance['title'] ) ) {
-						$output .= $args['before_title'] . apply_filters( 'widget_title', $instance['title'] ) . $args['after_title'];
+						$output .= $args['before_title'] . apply_filters( 'widget_title', $instance['title'], $instance, $this->id_base, $args ) . $args['after_title'];
 					}
 
 					$output .= '<div class="wm-twitter-container">';